requirements.txt
numpy
matplotlib
seaborn
tensorflow
keras
scikit-learn
imblearn
Pillow
opencv-python
"
